📝refactor #1 Reorganized components in atomic format.
