+++ components/atoms/CreditStats/index.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import React from "react"
 import Image from "next/image"
 import { Flex, Typography } from "antd"
