+++ components/index.tsx
@@ -1,2 +1,3 @@
-export * from "./MainLayout"
-export * from "./CreditStats"
+export * from "./templates"
+export * from "./atoms"
+export * from "./molecules"
