+++ components/templates/MainLayout/styles.ts
