+++ theme/themeConfig.ts
@@ -6,6 +6,18 @@ export const themeConfig: ThemeConfig["token"] = {
   fontFamily: "'Noto Sans JP', sans-serif",
 }
 
+export const customTheme = {
+  "gray-1": "#f4f5fc",
+  "gray-2": "#dce1ef",
+  "bg-main": "#F4F5FC",
+  fs: {
+    xsm: "12px",
+    sm: "14px",
+    md: "16px",
+    "2md": "20px",
+  },
+}
+
 export const antThemeConfig: ThemeConfig = {
   token: {
     ...colorScheme,
@@ -18,11 +30,13 @@ export const antThemeConfig: ThemeConfig = {
       controlHeight: 32,
       contentFontSize: 14,
       fontWeight: 500,
+      colorText: colorScheme?.["blue-10"],
     },
 
     Typography: {
       fontFamilyCode: "Roboto",
       fontSizeHeading5: 12,
+      fontSizeHeading4: 14,
       fontSize: 14,
       fontWeightStrong: 500,
       titleMarginBottom: 0,
@@ -31,19 +45,26 @@ export const antThemeConfig: ThemeConfig = {
       marginXXS: 0,
     },
 
+    Table: {
+      headerSplitColor: customTheme["gray-2"],
+      headerBg: customTheme["gray-2"],
+      headerColor: colorScheme?.["blue-10"],
+      fontSize: 14,
+      fontWeightStrong: 700,
+      cellFontSize: 14,
+      footerBg: "#fff",
+      stickyScrollBarBg: "#babac0",
+      stickyScrollBarBorderRadius: 80,
+      rowHoverBg: "#F4F5FC",
+      headerBorderRadius: 6,
+    },
+
     Layout: {
       siderBg: "#FFFFFF",
     },
-  },
-}
-
-export const customTheme = {
-  "gray-1": "#f4f5fc",
-  "bg-main": "#F4F5FC",
-  fs: {
-    xsm: "12px",
-    sm: "14px",
-    md: "16px",
-    "2md": "20px",
+    Switch: {
+      handleSize: 18,
+      trackHeight: 22,
+    },
   },
 }
